Bump minSdk to API 14
The Android v4 and v7 support libraries now only support a minSdk
of v14. This means that they are both essentially a v14 support
library now.

In order to bump compileSdk to 26, we'll have to update the support
library, which will break < v14 compatibility. Android is
essentially forcing all applications to break compatibility with
devices older than API 14.

// FREEBIE