Fix race condition that occurs on device link authorisation.
Upon authorisation, we upload our mapping to the server.
At the same time we also get a contact sync message and send out background friend request messages.
There was a race condition between those 2 functions where to correctly establish multi-device communication, you need your mapping on the server so that the other party knows that it's a secondary device and not a regular user.